// packages/api/src/api/types/generated.ts
export type ConfigMap = Record<string, string>;

/**
 * DevicesMap type is used to hold incus devices configurations. In contrast to
 * plain map[string]map[string]string it provides unmarshal methods for JSON and
 * YAML, which gracefully handle numbers and bools.
 *
 *
 * @example {"eth0": {"network": "incusbr0", "type": "nic"}}
 */
export type DevicesMap = Record<string, Record<string, string>>;
// ...
export interface Instance extends InstancePut {
  /**
   * Instance creation timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  created_at: string;
  /**
   * Expanded configuration (all profiles and local config merged)
   * @example {"security.nesting": "true"}
   */
  expanded_config?: ConfigMap;
  /**
   * Expanded devices (all profiles and local devices merged)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  expanded_devices?: DevicesMap;
  /**
   * Instance name
   * @example foo
   */
  name: string;
  /**
   * Instance status (see instance_state)
   * @example Running
   */
  status: string;
  /**
   * Instance status code (see instance_state)
   * @example 101
   */
  status_code: StatusCode;
  /**
   * Last start timestamp
   * @example 2021-03-23T20:00:00-04:00
   */
  last_used_at: string;
  /**
   * What cluster member this instance is located on
   * @example server01
   */
  location: string;
  /**
   * The type of instance (container or virtual-machine)
   * @example container
   */
  type: string;
  /**
   * Instance project name
   * @example foo
   *
   * API extension: instance_all_projects
   */
  project: string;
}
// ...
export interface InstancePut {
  /**
   * Architecture name
   * @example x86_64
   */
  architecture: string;
  /**
   * Instance configuration (see doc/instances.md)
   * @example {"security.nesting": "true"}
   */
  config: ConfigMap;
  /**
   * Instance devices (see doc/instances.md)
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}}
   */
  devices: DevicesMap;
  /**
   * Whether the instance is ephemeral (deleted on shutdown)
   * @example false
   */
  ephemeral: boolean;
  /**
   * List of profiles applied to the instance
   * @example ["default"]
   */
  profiles: string[];
  /**
   * If set, instance will be restored to the provided snapshot name
   * @example snap0
   */
  restore?: string;
  /**
   * Whether only the instances disk should be restored
   * @example false
   */
  disk_only?: boolean;
  /**
   * Whether the instance currently has saved state on disk
   * @example false
   */
  stateful: boolean;
  /**
   * Instance description
   * @example My test instance
   */
  description: string;
}
// ...
export type StatusCode = 100 | 101 | 102 | 103 | 104 | 105 | 106 | 107 | 108 | 109 | 110 | 111 | 112 | 113 | 200 | 400 | 401 | (number & {});